pyproject.toml
[project]
name = "jarvis_cd"
description = "Jarvis-CD: Unified platform for deploying applications and benchmarks"
readme = "README.md"
requires-python = ">=3.7"
dynamic = ["version"]
authors = [
  {name = "Luke Logan", email = "llogan@hawk.iit.edu"},
]
classifiers = [
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: System :: Systems Administration",
  "Topic :: System :: Distributed Computing",
]
dependencies = [
  "pyyaml",
  "pandas",
]

[project.urls]
Documentation = "https://github.com/iowarp/ppi-jarvis-cd"
issue-tracker = "https://github.com/iowarp/ppi-jarvis-cd/issues"
source-code = "https://github.com/iowarp/ppi-jarvis-cd"

[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools>=42",
  "setuptools-scm>=7",
]

[tool.setuptools]
packages = ["jarvis_cd", "jarvis_cd.basic", "jarvis_cd.core", "jarvis_cd.util", "builtin", "builtin.builtin"]
include-package-data = true

[tool.setuptools.package-data]
jarvis_cd = ["*.yaml", "*.yml"]
"*" = ["*.py", "*.yaml", "*.yml", "*.md", "*.txt", "*.conf", "*.xml", "*.param", "*.f", "*.input", "*.png", "*.sh"]

[project.scripts]
jarvis = "jarvis_cd.core.cli:main"

[tool.setuptools_scm]
version_scheme = "no-guess-dev"
